refactor: move header and footer components to core module. Add index.ts files for pretty imports.
