+++ src/app/about/about-routing.module.ts
@@ -1,10 +1,12 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { AboutPageComponent } from './pages/about-page/about-page.component';
-import { StyleguidePageComponent } from './pages/styleguide-page/styleguide-page.component';
-import { TeamPageComponent } from './pages/team-page/team-page.component';
-import { WorklogPageComponent } from './pages/worklog-page/worklog-page.component';
+import {
+  AboutPageComponent,
+  StyleguidePageComponent,
+  TeamPageComponent,
+  WorklogPageComponent
+} from '@about/pages';
 
 const routes: Routes = [
   {
