+++ src/app/about/components/index.ts
@@ -1,0 +1,4 @@
+export * from './difficulties/difficulties.component';
+export * from './evaluation/evaluation.component';
+export * from './team-member-card/team-member-card.component';
+export * from './worklog/worklog.component';
