+++ src/app/about/components/team-member-card/team-member-card.component.ts
@@ -1,4 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
+
 import { TeamMember } from '@about/models';
 
 @Component({
