+++ src/app/about/index.ts
@@ -1,0 +1,2 @@
+export * from './about.module';
+export * from './about-routing.module';
