+++ src/app/about/pages/index.ts
@@ -1,0 +1,4 @@
+export * from './about-page/about-page.component';
+export * from './styleguide-page/styleguide-page.component';
+export * from './team-page/team-page.component';
+export * from './worklog-page/worklog-page.component';
