+++ src/app/about/services/index.ts
@@ -1,0 +1,2 @@
+export * from './evaluation.service';
+export * from './team.service';
