+++ src/app/core/components/footer/footer.component.scss
