+++ src/app/core/components/index.ts
@@ -1,0 +1,2 @@
+export * from './header/header.component';
+export * from './footer/footer.component';
