+++ src/app/core/core.module.ts
@@ -1,13 +1,18 @@
 import { NgModule } from '@angular/core';
 
 import { SharedModule } from '@shared/shared.module';
-import { Error404PageComponent } from './pages/error404-page/error404-page.component';
-import { MainPageComponent } from './pages/main-page/main-page.component';
+import { Error404PageComponent, MainPageComponent } from '@core/pages';
 import { PhotographersModule } from '@photographers/photographers.module';
+import { FooterComponent, HeaderComponent } from '@core/components';
 
 @NgModule({
-  declarations: [Error404PageComponent, MainPageComponent],
+  declarations: [
+    Error404PageComponent,
+    MainPageComponent,
+    HeaderComponent,
+    FooterComponent
+  ],
   imports: [SharedModule, PhotographersModule],
-  exports: [MainPageComponent]
+  exports: [MainPageComponent, HeaderComponent, FooterComponent]
 })
 export class CoreModule {}
