+++ src/app/core/pages/index.ts
@@ -1,0 +1,2 @@
+export * from './error404-page/error404-page.component';
+export * from './main-page/main-page.component';
