+++ src/app/photographers/components/index.ts
@@ -1,0 +1,7 @@
+export * from './map/map.component';
+export * from './/photographer-card/photographer-card.component';
+export * from './photographer-item/photographer-item.component';
+export * from './search/search.component';
+export * from './timeline/timeline.component';
+export * from './videoplayer/videoplayer.component';
+export * from './worklist/worklist.component';
