+++ src/app/photographers/components/search/search.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { SearchService } from '@core/services/search.service';
+
+import { SearchService } from '@shared/services';
 
 @Component({
   selector: 'app-search',
