+++ src/app/photographers/components/timeline/timeline.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
+
 import { Event } from '@core/models';
 
 @Component({
