+++ src/app/photographers/pages/index.ts
@@ -1,0 +1,2 @@
+export * from './photographer-detail-page/photographer-detail-page.component';
+export * from './photographers-list-page/photographers-list-page.component';
