+++ src/app/photographers/photographers-routing.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { PhotographersListPageComponent } from './pages/photographers-list-page/photographers-list-page.component';
-import { PhotographerDetailPageComponent } from './pages/photographer-detail-page/photographer-detail-page.component';
-import { PhotographerResolver } from '@photographers/services/photographer.resolver';
+import { PhotographersListPageComponent } from '@photographers/pages';
+import { PhotographerDetailPageComponent } from '@photographers/pages';
+import { PhotographerResolver } from '@photographers/services';
 
 const routes: Routes = [
   { path: '', component: PhotographersListPageComponent },
