+++ src/app/photographers/photographers.module.ts
@@ -1,26 +1,23 @@
 import { NgModule } from '@angular/core';
 
-import { SharedModule } from '@shared/shared.module';
-import { PhotographersRoutingModule } from './photographers-routing.module';
 import { MglTimelineModule } from 'angular-mgl-timeline.9';
 import { AngularYandexMapsModule } from 'angular8-yandex-maps';
 import { YouTubePlayerModule } from '@angular/youtube-player';
 import { NgImageSliderModule } from 'ng-image-slider';
 
-import { PhotographersListPageComponent } from './pages/photographers-list-page/photographers-list-page.component';
-import { PhotographerDetailPageComponent } from './pages/photographer-detail-page/photographer-detail-page.component';
-import { PhotographerItemComponent } from './components/photographer-item/photographer-item.component';
-import { TimelineComponent } from './components/timeline/timeline.component';
-import { SearchComponent } from './components/search/search.component';
-import { PhotographerCardComponent } from './components/photographer-card/photographer-card.component';
-import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
-import { FbTransLoaderFactory } from '../firebase.module';
-import { AngularFireDatabase } from 'angularfire2/database';
-import { MapComponent } from './components/map/map.component';
-import { WorklistComponent } from './components/worklist/worklist.component';
-import { VideoplayerComponent } from './components/videoplayer/videoplayer.component';
 import { environment } from 'src/environments/environment';
-import { SearchPipe } from '@core/pipes/search.pipe';
+import { PhotographersRoutingModule } from './photographers-routing.module';
+import { SharedModule } from '@shared/shared.module';
+
+import { PhotographersListPageComponent } from '@photographers/pages';
+import { PhotographerDetailPageComponent } from '@photographers/pages';
+import { PhotographerItemComponent } from '@photographers/components';
+import { TimelineComponent } from '@photographers/components';
+import { SearchComponent } from '@photographers/components';
+import { PhotographerCardComponent } from '@photographers/components';
+import { MapComponent } from '@photographers/components';
+import { WorklistComponent } from '@photographers/components';
+import { VideoplayerComponent } from '@photographers/components';
 
 @NgModule({
   declarations: [
@@ -32,8 +29,7 @@ import { SearchPipe } from '@core/pipes/search.pipe';
     PhotographerCardComponent,
     MapComponent,
     WorklistComponent,
-    VideoplayerComponent,
-    SearchPipe
+    VideoplayerComponent
   ],
   imports: [
     SharedModule,
@@ -41,15 +37,7 @@ import { SearchPipe } from '@core/pipes/search.pipe';
     MglTimelineModule,
     YouTubePlayerModule,
     NgImageSliderModule,
-    AngularYandexMapsModule.forRoot(environment.ymapsApiKey),
-    TranslateModule.forChild({
-      loader: {
-        provide: TranslateLoader,
-        useFactory: FbTransLoaderFactory,
-        deps: [AngularFireDatabase]
-      },
-      extend: true
-    })
+    AngularYandexMapsModule.forRoot(environment.ymapsApiKey)
   ],
   exports: [PhotographerItemComponent]
 })
