+++ src/app/photographers/services/index.ts
@@ -1,0 +1,2 @@
+export * from './photographer.resolver';
+export * from './photographer.service';
