+++ src/app/shared/firestore-trans-loader.ts
@@ -1,5 +1,5 @@
-import { TranslateLoader } from '@ngx-translate/core';
 import { AngularFireDatabase } from 'angularfire2/database';
+import { TranslateLoader } from '@ngx-translate/core';
 import { Observable } from 'rxjs';
 
 export class FirestoreTransLoader implements TranslateLoader {
