+++ src/app/shared/pipes/index.ts
@@ -1,0 +1,2 @@
+export * from './date-translate.pipe';
+export * from './search.pipe';
