+++ src/app/shared/pipes/search.pipe.ts
@@ -1,4 +1,5 @@
 import { Pipe, PipeTransform } from '@angular/core';
+
 import { Author } from '@core/models';
 
 @Pipe({
