+++ src/app/shared/services/index.ts
@@ -1,1 +1,2 @@
 export * from './language-settings.service';
+export * from './search.service';
