+++ src/app/shared/shared.module.ts
@@ -2,23 +2,20 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
 
-import { HeaderComponent } from './components/header/header.component';
-import { FooterComponent } from './components/footer/footer.component';
-import { AngularMaterialModule } from './angular-material.module';
-
 import { TranslateModule } from '@ngx-translate/core';
-import { DateTranslatePipe } from './pipes/date-translate.pipe';
+import { AngularMaterialModule } from '@shared/angular-material.module';
+import { DateTranslatePipe, SearchPipe } from '@shared/pipes';
 
 @NgModule({
-  declarations: [HeaderComponent, DateTranslatePipe, FooterComponent],
+  declarations: [DateTranslatePipe, SearchPipe],
   imports: [CommonModule, TranslateModule, RouterModule, AngularMaterialModule],
   exports: [
     CommonModule,
+    RouterModule,
     TranslateModule,
-    HeaderComponent,
     AngularMaterialModule,
     DateTranslatePipe,
-    FooterComponent
+    SearchPipe
   ]
 })
 export class SharedModule {}
